Fix Naive Else Bug (#15)
* Fix bug with a To not sharing its Else

The original solution for the Else was naive in that there wasn't a single Else shared amongst the To's
Have and Be fields.  Calls like expect(foo).To.Equal(bar).And.Be.String().Else.FailNow() would
therefor not call FailNow() on the *testing.T type.

The solution was simply to pass the same *Else value along to the To's Have and Be fields.

* Unexport NewTo et. al. before we get ourselves into trouble

Supporting those constructors as part of our public API is going to get really nasty, especially
as we're not currently accessing them directly in our tests.